Move the product select to its own component
