+++ packages/js/product-editor/src/blocks/generic/linked-product-list/editor.scss
@@ -2,51 +2,4 @@
 	display: flex;
 	flex-direction: column;
 	gap: $grid-unit-30;
-
-	&__menu {
-		&-loading {
-			display: flex;
-			align-items: center;
-			justify-content: center;
-			padding: $grid-unit;
-			height: $grid-unit-60 + 2px;
-		}
-
-		&-item {
-			display: flex;
-			align-items: center;
-			padding: $grid-unit;
-			gap: $grid-unit + $grid-unit-05;
-
-			&-image {
-				width: $grid-unit-40;
-				height: $grid-unit-40;
-				border-radius: $grid-unit-05;
-				background-color: $gray-200;
-				background-position: center center;
-				background-repeat: no-repeat;
-				background-size: cover;
-				flex-shrink: 0;
-			}
-
-			&-content {
-				display: flex;
-				flex-direction: column;
-				gap: 2px;
-			}
-
-			&-title {
-				color: $gray-900;
-			}
-
-			&-description {
-				color: $gray-700;
-				font-size: 11px;
-			}
-		}
-	}
-
-	.woocommerce-experimental-select-control__combo-box-wrapper {
-		margin-bottom: 0;
-	}
 }
