+++ packages/js/product-editor/src/components/product-select/index.ts
@@ -1,0 +1,2 @@
+export * from './product-select';
+export * from './types';
