+++ packages/js/product-editor/src/style.scss
@@ -43,6 +43,7 @@
 @import "components/button-with-dropdown-menu/style.scss";
 @import "components/formatted-price/style.scss";
 @import "components/product-list/styles.scss";
+@import "components/product-select/styles.scss";
 
 /* Field Blocks */
